Multiplayer testing... not going very great
BUILD: Bump version

So, I tested the mod being installed on a server. It didn't seem to do anything. However, I also tested the mod without being installed on the server, but installed on the client. It worked...???????????????? It shouldn't work, as it could be an advantage to the client which has it installed. I have no idea how to fix this

Signed-off-by: Superkat32 <superkat1806@gmail.com>