Multiplayer tested and version bump
BUILD: Bump version

I have tested the mod on a multiplayer server outside of dev env. It works much better than last time(crashed). There are still some QoL things I would like to add, such as the ability to disable the dust particles client-side. And to allow the dust particles to work client-side should the mod not be installed on the server(probably have to rework some stuff for that specific environment).

Signed-off-by: Superkat32 <superkat1806@gmail.com>