Removed speed/jump buffs
BUILD: Bump version
REFACTOR: Removed the speed buff
REFACTOR: Removed the jump buff

I'm REALLY struggling to have this work correctly when used on a server(both server and client side). The issue is that the client is assuming things, and messing with stuff it shouldn't. There also doesn't seem to be an easy way to detect from the client side if the server has a mod installed or not. So, this will be one of many moves made to help the mod work correctly when installed on the client, but not the server.

Signed-off-by: Superkat32 <superkat1806@gmail.com>